change MPI interceptor language from C to C++
